requirements.txt
pandas==0.25.3
streamlit==0.67.0
numpy==1.18.5
python_bidi==0.4.2
plotly==4.8.1
matplotlib==3.2.1
wordcloud==1.7.0
scikit_learn==0.23.2
